initialize development
