+++ package.json
@@ -1,0 +1,34 @@
+{
+  "name": "zesty.io-fetch-wrapper",
+  "version": "0.0.1",
+  "description": "",
+  "main": "dist/index_bundle.js",
+  "scripts": {
+    "build": "webpack",
+    "release": "release-it"
+  },
+  "keywords": [],
+  "author": "Hridayesh Sharma",
+  "license": "ISC",
+  "dependencies": {},
+  "devDependencies": {
+    "@babel/core": "^7.10.4",
+    "@babel/preset-env": "^7.11.0",
+    "babel-core": "^6.26.3",
+    "babel-loader": "^8.2.5",
+    "babel-polyfill": "^6.26.0",
+    "babel-preset-es2015": "^6.24.1",
+    "babel-preset-stage-0": "^6.24.1",
+    "release-it": "^14.14.2",
+    "webpack": "^4.44.1",
+    "webpack-cli": "^3.3.12"
+  },
+  "browserslist": [
+    "last 2 Chrome versions"
+  ],
+  "release-it": {
+    "github": {
+      "release": true
+    }
+  }
+}
