+++ src/index.js
@@ -1,0 +1,3 @@
+import ZestyioRestFetchWrapper from "./fetchApi";
+
+export { ZestyioRestFetchWrapper };
