+++ test.html
@@ -1,0 +1,33 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>Demo</title>
+  </head>
+  <body>
+    <script src="/dist/index_bundle.js"></script>
+    <script>
+      function getCookie(cname) {
+        var name = cname + "=";
+        var ca = document.cookie.split(";");
+        for (var i = 0; i < ca.length; i++) {
+          var c = ca[i];
+          while (c.charAt(0) == " ") {
+            c = c.substring(1);
+          }
+          if (c.indexOf(name) == 0) {
+            return c.substring(name.length, c.length);
+          }
+        }
+        return "";
+      }
+      console.log($);
+      ZestyAPI = new $.ZestyioRestFetchWrapper(
+        "8-82f8f3fff6-4zq0d0",
+        getCookie("APP_SID")
+      );
+      console.log(ZestyAPI);
+    </script>
+  </body>
+</html>
