+++ webpack.config.js
@@ -1,0 +1,21 @@
+const path = require("path");
+
+module.exports = {
+  entry: path.resolve(__dirname, "src/index.js"),
+  output: {
+    path: path.resolve(__dirname, "dist"),
+    filename: "index_bundle.js",
+    library: "$",
+    libraryTarget: "umd",
+  },
+  module: {
+    rules: [
+      {
+        test: /\.(js)$/,
+        exclude: /node_modules/,
+        use: "babel-loader",
+      },
+    ],
+  },
+  mode: "development",
+};
